chore: clean up test code
